Sample Scene made in Special Room
