+++ Assets/Scripts/SceneScripts/SampleScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2aa903c35d3250247a76744fb9204c67
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
